Adds detection of conflicting operation ids to code gen
Adding this led to the discovery that we were not propertly handling
fallback name inference for operations with different methods on the
same path item. This is now fixed -- the method name is used to
disambiguate the operations modules when multple methods can be used on
the same endpoint and no operation ids are given the OpenAPI
description.