Adds a pretty decoder that removes Aeson portion of decoding error
This splits the error message produced by `fromValue` into two:
the Aeson and Fleece error message combined (the original behavior)
and just the Fleece error message (that produced by the schema and
the `Decoder` functions). `decodePretty`, as a result, produces a
slightly easier to understand error message that is more suitable
for displaying to users. Ideally, we would be able to remove the
type from the error messages produced by the `Decoder` functions as
well, but for now, this isn't changing that since it would alter
the current tests' expected results.