Adds mechanism for converting Fleece schemas to Aeson Values
This allows for the use of Fleece schemas in defining `ToJSON`
instances, like we do using `toParser` for `FromJSON` instances.

Originally the plan was to replace the existing `Encoder` using
`Encoding` and `Series` with one that uses `Value` and adding the
`toEncoding` step in `encode` and `encodeStrict`. This results in
a different JSON encoding, however - insertion order for key/value
pairs is not preserved when using `Value`. This resulted in test
failures because while the same fields are present, they may appear
in a different order, so the encoded bytestrings won't always
match.