Sorts fields in records generated via OpenApi
Although we would like for the fields in generated records to match the
order that they're specified in the OpenApi spec, library support for
preserving the order while parsing the `OpenApi` description is currently
lacking. In order to provide at least a *stable* order that does not vary
with changes to the underlying libraries, this sorts the fields by their
name while constructing the description of the object for the code
generation layer.