Sets up CI and local builds for supported GHC versions
This add's a matrix build for GHC versions 9.4 through 9.10, along
with a couple of minor code fixes to make `json-fleece` work across
these versions.

This uses a new strategy for defining the `stack-ghc*.yaml` files
that utilizes the `!include` yaml syntax that our `stack` fork now
supports (and we will upstream). This allows us to define the
stack files with a minimal amount of duplication. This also makes it
feasible to have `stack.yaml` to effectively act as a passthrough to
whatever ghc version-specific yaml file we want to use for regular
development.

The build updates here attempt to be fairly precise about the github
actions caches. Each build job gets its own cache since they will be
caching different GHC compilers and dependencies. This should make
each cache smaller while making it less likely that they will trample
each other in the matrix build. The restore keys have been set
accordingly. The full cache key includes the hash of ann the stack.yaml
files and package files that could impact that particular build to
try to save on pushing new caches where possible while still creating
new ones when it is beneficial to do so.