Adds inline object support inside of oneOf
Naming inline objects was the biggest roadblock to supporting this,
and our solution was to name each schema within the oneOf as
"OptionN", where N is that schema's index in the list of oneOf
options.

We also addressed nested arrays, where we append "Item" to the
end of the schema name. This gets duplicated for each level of
nesting. We also did some work to make sure that the "OptionN"
part for inline objects inside oneOf to come before the "Item"s.

Finally, we fixed inline strings in oneOfs that represent `Enum`s.
This was also fixed by appending "OptionN" to the end so they don't
cause a name conflict.

Co-authored-by: David <david@flipstone.com>