Inline allOf everywhere before codegen
Processing and de-referencing schemas listed in "allOf" is now a top-level
transformation done prior to codegen. This ensures that "allOf" is
handled consistently across responses, parameters, and anywhere else a
schema shows up.

The new "AllOfObject" test case is meant to highlight the behavior and
expected output of this pre-procesing step.

Machinery for doing this is provided in OpenApi3.Traversal, which allows us
to traverse every part of an OA.OpenApi that contains a referenced schema,
even those we do not currently support.

Co-authored-by: Tyler Baum <tyler@flipstone.com>
Co-authored-by: Nebula Juniper Lavelle <nebula@flipstone.com>
Co-authored-by: Janus <janus@flipstone.com>