Inline allOf everywhere before codegen
Schemas inside `allOf` arrays will be merged into their parent
schemas as a pre-processing step before proceeding with codegen.

We apply the transformation to every 'OA.Referenced OA.Schema' in the
'OA.OpenApi' via a new traversal function defined in
`Fleece.OpenApi3.Traversal`.

We added an "AllOfObject" test case is meant to highlight the behavior
and expected output of this pre-processing step.

Co-authored-by: Tyler Baum <tyler@flipstone.com>
Co-authored-by: Nebula Juniper Lavelle <nebula@flipstone.com>
Co-authored-by: Janus <janus@flipstone.com>