Merge identically named properties in allOf correctly
In a previous commit, we rejected properties in `allOf`s with identical
names. This was too aggressive and prevented us from parsing some valid
schemas. Now, we attempt to merge these schemas so long as both are
inline. We still reject identically named schemas when either or both
are referenced.

Co-authored-by: Tyler Baum <tyler@flipstone.com>