Adds Fleece API members to aid in generating better names in OpenAPI schemas
Adds `validateAnonymous` and a number of associated functions
---
These functions allow you to define a schema that transforms the
Haskell representation of schema without altering the name that is
used to refer to the type when referencing the json values.

Some schema helpers such as `list` where changed to use this to make
them transparent to the underlying array type in JSON.

Adds `format`
---
This allows schema constructors to specify the format of the value that
is expected for a schema. The ISO8601 time helpers now use
`validateAnonymous` and `format` to cause allow OpenAPI specs to be
generated that simply show a type of `string` with the appropriate
`format` specified.

Removes bareJSONOrString
---
This function is problematic and is not actually used in any of our
code anymore, so I removed it.

Change `int`, `int8`, etc to be class members
---
These members have defaults which use `validateAnonymous` and `format`
to specify consistent encoding/decoding behavior for implementations
that don't override them. The are now class members so that implementations
may override them if they need to provide more specific behavior for
these types beyond specifying the `format`.

Converts markdown tests to golden tests
---
These tests are easier to read, digest and maintain this way