Universally quanitifies schema arguments in Fleece methods
All of the `Fleece` methods that take a schema as an argument now have
those schemas universally quantified, i.e. all `schema a` arguments are
now `forall anySchema. Fleece anySchema => anySchema a`. This allows
instances to interpret the schemas using another instance.

The original motivation behind this change was to provide a canonical
way to derive the name of a schema that was not dependent on the
concrete instance. Previously it was up to the instance owners to ensure
that the result of `schemaName` was consistent with other instances.

This commit removes the `schemaName` method from `Fleece` and instead
introduces a new `SchemaName` instance with a `schemaName :: Name`
field. In conjuction with the universal quantification of the schema
arguments, this allows access to the canonical name for any universally
quantified schema.